+++ global.d.ts
@@ -147,8 +147,8 @@ interface TeamForClub {
 
 interface DetailedMatchInfo extends Match {
   location: Location
-  previousEncounters: Match[]
   strengthDifference: number | null
   strengthDifferenceWithoutCurrent: number | null
   fullTeamName: string
+  otherEncounters: Match[]
 }